update editar usuario

+++ assets/php/usuarios/eliminarUsuario.php
@@ -2,28 +2,19 @@
 include("../Conexion.php");
 $db = new Conexion();
 
-$cedula = $_GET['cedula'];
-
+$bandera = false;
 echo("
    <script>
-   function ConfirmDemo() {
-      //Ingresamos un mensaje
-      var mensaje = confirm('¿Estás seguro que quieres eliminar esta cuenta?');
-      //Verificamos si el usuario acepto el mensaje
-      if (mensaje) {".eliminarUsuario()."
-      }
-      //Verificamos si el usuario denegó el mensaje
-      else {
-      alert('¡Haz denegado la eliminación de la cuenta');
-      }
-   }
+         var mensaje = confirm('¿Estás seguro que quieres eliminar esta cuenta?');
    </script>
-
 ");
 
+if(isset($_GET['mensaje'])){
+   $bandera= $_GET['mensaje'];
+} 
 
-function eliminarUsuario(){
-   global $cedula;
+if($bandera){
+   $cedula = $_GET['cedula'];
    $sql = "DELETE FROM `cliente` WHERE cedula LIKE '$cedula'";
    $result = $db->OperSql($sql);
 
@@ -37,6 +28,11 @@ function eliminarUsuario(){
    } else {
       http_response_code(500);
    }
+}else{
+   echo ("
+      <script>
+      alert('Gil');
+      window.location = '../../../editarUsuario.php';
+      </script>'");
 }
-
 ?>
